Созданы раздельные модели под задачи программы.
